+++ trikot-viewmodels-declarative/compose/src/main/kotlin/com/mirego/trikot/viewmodels/declarative/compose/extensions/AnimationExtensions.kt
@@ -14,12 +14,11 @@ import com.mirego.trikot.viewmodels.declarative.animation.VMDAnimation
 import com.mirego.trikot.viewmodels.declarative.animation.VMDAnimationEasing
 import kotlin.time.DurationUnit
 
-fun <T, V> VMDAnimatedPropertyChange<T, V>.animationSpec() : AnimationSpec<T> {
-    return propertyChange.animation?.toComposeAnimationSpec() ?: snap()
-}
+fun <T, V> VMDAnimatedPropertyChange<T, V>.animationSpec() : AnimationSpec<T>  =
+    propertyChange.animation?.toComposeAnimationSpec() ?: snap()
 
-fun <T> VMDAnimation.toComposeAnimationSpec() : AnimationSpec<T> {
-    return when (this) {
+fun <T> VMDAnimation.toComposeAnimationSpec() : AnimationSpec<T> =
+    when (this) {
         is VMDAnimation.Tween -> {
             tween(
                 durationMillis = duration.toInt(DurationUnit.MILLISECONDS),
@@ -31,10 +30,9 @@ fun <T> VMDAnimation.toComposeAnimationSpec() : AnimationSpec<T> {
             spring(dampingRatio = dampingRatio, stiffness = stiffness)
         }
     }
-}
 
-fun VMDAnimationEasing.toComposeEasing() : Easing {
-    return when (this) {
+fun VMDAnimationEasing.toComposeEasing() : Easing =
+    when (this) {
         VMDAnimationEasing.Standard.Linear -> LinearEasing
         VMDAnimationEasing.Standard.EaseIn -> LinearOutSlowInEasing
         VMDAnimationEasing.Standard.EaseOut -> FastOutLinearInEasing
@@ -43,4 +41,3 @@ fun VMDAnimationEasing.toComposeEasing() : Easing {
             CubicBezierEasing(a = a, b = b, c = c, d = d)
         }
     }
-}
